Create Console Message Helper class
